Admin 테이블 및 패키지 삭제 하고 login으로 옮기면서 정리하는 과정 (에러 상황)
