--- src/main/java/com/acorn/escape/admin/dao/AdminDao.java
@@ -1,7 +1,0 @@
-package com.acorn.escape.admin.dao;
-
-import com.acorn.escape.admin.dto.AdminDto;
-
-public interface AdminDao {
-	public boolean isValid(AdminDto dto);
-}
