--- src/main/java/com/acorn/escape/admin/service/AdminService.java
@@ -1,12 +1,0 @@
-package com.acorn.escape.admin.service;
-
-import javax.servlet.http.HttpSession;
-
-import org.springframework.web.servlet.ModelAndView;
-
-import com.acorn.escape.admin.dto.AdminDto;
-
-public interface AdminService {
-	public void loginProcess(AdminDto dto, ModelAndView mView,
-			HttpSession session);
-}
