+++ src/main/java/com/acorn/escape/mybatis/Configuration.xml
@@ -6,9 +6,10 @@
 	<typeAliases>
 		<typeAlias type="com.acorn.escape.notice.dto.NoticeDto"
 			alias="noticeDto"/>
-		<typeAlias type="com.acorn.escape.login.dto.LoginDto" alias="loginDto"/>
+		<typeAlias type="com.acorn.escape.login.dto.LoginDto" 
+			alias="loginDto"/>
 	</typeAliases>
-	<!-- Mapper 등록 -->
+
 	<mappers>
 		<mapper resource="com/acorn/escape/mybatis/LoginMapper.xml"/>
 		<mapper resource="com/acorn/escape/mybatis/NoticeMapper.xml"/>
