+++ src/main/webapp/WEB-INF/views/notice/hongdae/list.jsp
@@ -34,10 +34,10 @@
 	<h4>공지사항 - 홍대점</h4>
 	<c:choose>
 		<c:when test="${empty aid }">
-			<a href="${pageContext.request.contextPath }/admin/loginform.do">로그인</a>
+			<a href="${pageContext.request.contextPath }/login/login_form.do">로그인</a>
 		</c:when>
 		<c:otherwise>
-			<a href="${pageContext.request.contextPath }/admin/logout.do">로그아웃</a>
+			<a href="${pageContext.request.contextPath }/login/logout.do">로그아웃</a>
 		</c:otherwise>
 	</c:choose>
 	<br/>
